put interface beside usage
