+++ packages/insomnia/src/ui/components/websockets/websocket-response-pane.tsx
@@ -7,9 +7,8 @@ import { ResponseTimelineEntry } from '../../../main/network/libcurl-promise';
 import { WebsocketEvent } from '../../../main/network/websocket';
 import * as models from '../../../models';
 import type { Response } from '../../../models/response';
-import { createWebSocketClient } from '../../context/websocket-client/create-websocket-client';
 import { useWebSocketConnectionEvents } from '../../context/websocket-client/use-ws-connection-events';
-import { WebSocketClientProvider } from '../../context/websocket-client/websocket-client-context';
+import { createWebSocketClient, WebSocketClientProvider } from '../../context/websocket-client/websocket-client-context';
 import { selectActiveResponse } from '../../redux/selectors';
 import { ResponseHistoryDropdown } from '../dropdowns/response-history-dropdown';
 import { Pane, PaneHeader as OriginalPaneHeader } from '../panes/pane';
