--- packages/insomnia/src/ui/context/websocket-client/create-websocket-client.ts
@@ -1,18 +1,0 @@
-export interface WebSocketClient {
-    event: typeof window.main.webSocketConnection.event;
-    create: typeof window.main.webSocketConnection.create;
-    close: typeof window.main.webSocketConnection.close;
-    send: typeof window.main.webSocketConnection.event.send;
-    onReadyState: typeof window.main.webSocketConnection.readyState.subscribe;
-    getReadyState: typeof window.main.webSocketConnection.readyState.getCurrent;
-  }
-export function createWebSocketClient(): WebSocketClient {
-  return {
-    event: window.main.webSocketConnection.event,
-    create: window.main.webSocketConnection.create,
-    close: window.main.webSocketConnection.close,
-    send: window.main.webSocketConnection.event.send,
-    onReadyState: window.main.webSocketConnection.readyState.subscribe,
-    getReadyState: window.main.webSocketConnection.readyState.getCurrent,
-  };
-}
