+++ packages/insomnia/src/ui/context/websocket-client/websocket-client-context.tsx
@@ -1,7 +1,5 @@
 import React, { createContext, FC, ReactNode, useContext } from 'react';
 
-import { WebSocketClient } from './create-websocket-client';
-
 const WebSocketClientContext = createContext<WebSocketClient | undefined>(undefined);
 interface Props {
   client: WebSocketClient;
@@ -26,4 +24,22 @@ export function useWebSocketClient(): WebSocketClient {
   }
 
   return context;
+}
+export interface WebSocketClient {
+  event: typeof window.main.webSocketConnection.event;
+  create: typeof window.main.webSocketConnection.create;
+  close: typeof window.main.webSocketConnection.close;
+  send: typeof window.main.webSocketConnection.event.send;
+  onReadyState: typeof window.main.webSocketConnection.readyState.subscribe;
+  getReadyState: typeof window.main.webSocketConnection.readyState.getCurrent;
+}
+export function createWebSocketClient(): WebSocketClient {
+  return {
+    event: window.main.webSocketConnection.event,
+    create: window.main.webSocketConnection.create,
+    close: window.main.webSocketConnection.close,
+    send: window.main.webSocketConnection.event.send,
+    onReadyState: window.main.webSocketConnection.readyState.subscribe,
+    getReadyState: window.main.webSocketConnection.readyState.getCurrent,
+  };
 }
